Fix calculation for multiple skill levels
